Enable lit testing for Android libc++ builds
The test suite uses "adb ..." to run tests, so ANDROID_SERIAL may need
to be set.

Set LIBCXX[ABI]_TEST_CONFIG and LIBCXX[ABI]_EXECUTOR.

Ensure that each libc++ test program is built with --target and
--sysroot:
 * Set CMAKE_{C,CXX}_COMPILER_TARGET instead of
   LIBCXX[ABI]_TARGET_TRIPLE. The LIBCXX[ABI]_TARGET_TRIPLE variables
   are no longer supported.
 * Set CMAKE_SYSROOT.

Also change the arm32 triple to start with armv7a. The libc++ tests
aren't compiled with -march=armv7-a, so we need the triple to start
with armv7a so that atomic tests pass. Also add the "eabi" suffix
because that's the triple that the NDK's arm32 wrappers use.

Bug: http://b/175635923
Test: build.py, then run the libc++ tests using lit manually
Change-Id: I4a2efd68db38e8ba5145dbf3db35fe0df1aeb6ad