Adding documentation
Starting with initial set of things that we like to share with
developers. We regularly find ourselves telling same things over and
over. Having a documentation will help us point to the link rather than
explain everytime. There is also a TODO list for those looking to
improve the doc overtime.

Change-Id: I2f9e920ae68bbfa3445f78e96be490eab070e0da