Ship a static libc++ build with exceptions off
The Bionic loader can't use ELF TLS itself, and libc++abi uses ELF TLS
for the exception handling globals. To avoid linking in TLS variables
into the loader, it needs to link a version of libc++ that has
exceptions turned off.

The archives are named:
 - libc++_static_noexcept.a
 - libc++demangle_noexcept.a
 - libc++abi_noexcept.a

The archives are linked into Bionic using special noexcept-variant
Soong modules, and because Soong requires prebuilt modules to have
unique filenames, append a "noexcept" suffix to the filenames.

Bug: http://b/175635923
Test: llvm_build/build.py, manually verify that linker[64] has no
  PT_TLS segment

Change-Id: I3499fff89df178f0320f566dcde14a472cfde0e0