LibInfo: omit version from link_libraries
Currently, we pass an absolute path to libxml2 to CMake, e.g.:

    -DLIBXML2_LIBRARY=/.../libxml2-linux-install/lib/libxml2.so.2

Instead, pass this:

    -DLIBXML2_LIBRARY=/.../libxml2-linux-install/lib/libxml2.so

install_libraries is unaffected: libxml2.so.2 is installed and
libxml2.so isn't.

LLVM's get_library_name CMake function strips prefixes (e.g. `lib`) and
suffixes (`.so`), but it doesn't/can't strip a version number, so it
turns the path above into `xml2.so.2`. This string is then used to
set LLVM_SYSTEM_LIBS on the LLVMWindowsManifest target, which then
propagates into `llvm-config --system-libs` as `-lxml2.so.2`, which is
an invalid linker command. It needs to be `-lxml2` instead.

On Darwin, either `-lxml2` or `-lxml2.2` works, but use `-lxml2` for
consistency with Linux.

Fix regex syntax used to parse `objdump -p`.

Bug: none
Test: build.py
Test: examine `llvm-config --system-libs` manually
Test: verify that `ldd` and `readelf -d` show the same dependencies
  across various binaries, like clang-19, lldb, ld.lld
Change-Id: I6f0748554f3e441bb63ff7366f473f39f3feb998