Add testcase for mapfile
It is quite confusing as not much documentation is there.

Not sure if this is the right directory to put the mapfile.
The test can be invoked simply like python /path/to/test_mapfile.py
Fixes maybe required if this file is moved elsewhere as it tries to find ndk releases
for using the libclang_rt.*so files for testing

Change-Id: I23d6699532553f0a77cc0ddbbbc2f25e06c2b9de